pyproject.toml
[build-system]
requires = ["setuptools>=67.0"]
build-backend = "setuptools.build_meta"

[project]
name = "translinguistic-nlp"
version = "0.1.0"
description = "Translinguistic knowledge graph pipeline with GNN-based link prediction"
authors = [{ name = "Translinguistic NLP" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pyyaml>=6.0",
    "requests>=2.32",
    "transformers>=4.44",
    "spacy>=3.7",
    "sentence-transformers>=3.0",
    "scikit-learn>=1.4",
    "scipy>=1.12",
    "neo4j>=5.18",
    "torch>=2.2",
    "torch-geometric>=2.5",
    "pandas>=2.2",
    "altair>=5.2",
    "beautifulsoup4>=4.12",
    "streamlit>=1.39",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
where = ["src"]
